refactor: 도메인명 수정 - generation -> recruitment
