--- src/main/java/com/bee/beehomepagebackend/generation/admin/question/controller/response/UpdateQuestionResponse.java
@@ -1,28 +1,0 @@
-package com.bee.beehomepagebackend.generation.admin.question.controller.response;
-
-import com.bee.beehomepagebackend.generation.question.domain.Question;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class UpdateQuestionResponse {
-
-    private Long id;
-    private Integer sequence;
-    private String content;
-
-    @Builder
-    public UpdateQuestionResponse(Long id, Integer sequence, String content) {
-        this.id = id;
-        this.sequence = sequence;
-        this.content = content;
-    }
-
-    public static UpdateQuestionResponse of(Question question) {
-        return UpdateQuestionResponse.builder()
-                .id(question.getId())
-                .sequence(question.getSequence())
-                .content(question.getContent())
-                .build();
-    }
-}
