--- src/main/java/com/bee/beehomepagebackend/generation/answer/domain/Answer.java
@@ -1,37 +1,0 @@
-package com.bee.beehomepagebackend.generation.answer.domain;
-
-import com.bee.beehomepagebackend.generation.applicant.domain.Applicant;
-import com.bee.beehomepagebackend.generation.question.domain.Question;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor
-@Entity
-@Table(name = "answer")
-public class Answer {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    private Applicant applicant;
-
-    @OneToOne
-    private Question question;
-
-    private String content;
-
-    @CreatedDate
-    private LocalDateTime createdDateTime;
-
-    @LastModifiedDate
-    private LocalDateTime modifiedDateTime;
-
-}
