--- src/main/java/com/bee/beehomepagebackend/generation/answer/domain/AnswerRepository.java
@@ -1,9 +1,0 @@
-package com.bee.beehomepagebackend.generation.answer.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AnswerRepository extends JpaRepository<Answer, Long> {
-
-}
