--- src/main/java/com/bee/beehomepagebackend/generation/applicant/controller/request/CreateApplicantRequest.java
@@ -1,9 +1,0 @@
-package com.bee.beehomepagebackend.generation.applicant.controller.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class CreateApplicantRequest {
-}
