--- src/main/java/com/bee/beehomepagebackend/generation/applicant/controller/response/CreateApplicantResponse.java
@@ -1,11 +1,0 @@
-package com.bee.beehomepagebackend.generation.applicant.controller.response;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class CreateApplicantResponse {
-
-
-}
