--- src/main/java/com/bee/beehomepagebackend/generation/applicant/domain/ApplicantRepository.java
@@ -1,8 +1,0 @@
-package com.bee.beehomepagebackend.generation.applicant.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ApplicantRepository extends JpaRepository<Applicant, Long> {
-}
