--- src/main/java/com/bee/beehomepagebackend/generation/applicant/enums/Position.java
@@ -1,9 +1,0 @@
-package com.bee.beehomepagebackend.generation.applicant.enums;
-
-public enum Position {
-
-    BACKEND_DEVELOPER,
-    FRONTEND_DEVELOPER,
-    DESIGNER;
-
-}
