--- src/main/java/com/bee/beehomepagebackend/generation/applicant/service/ApplicantService.java
@@ -1,4 +1,0 @@
-package com.bee.beehomepagebackend.generation.applicant.service;
-
-public class ApplicantService {
-}
