--- src/main/java/com/bee/beehomepagebackend/generation/applicant/service/request/CreateApplicantServiceRequest.java
@@ -1,9 +1,0 @@
-package com.bee.beehomepagebackend.generation.applicant.service.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class CreateApplicantServiceRequest {
-}
