--- src/main/java/com/bee/beehomepagebackend/generation/question/controller/QuestionController.java
@@ -1,25 +1,0 @@
-package com.bee.beehomepagebackend.generation.question.controller;
-
-import com.bee.beehomepagebackend.generation.question.service.QuestionService;
-import com.bee.beehomepagebackend.generation.question.response.GetAllQuestionsResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RestController
-@RequestMapping("/api/generations/questions")
-@RequiredArgsConstructor
-public class QuestionController {
-
-    private final QuestionService questionService;
-
-    @GetMapping
-    public ResponseEntity<GetAllQuestionsResponse> getAllQuestions() {
-        return ResponseEntity.ok(questionService.getAllQuestions());
-    }
-
-}
