--- src/main/java/com/bee/beehomepagebackend/generation/question/domain/QuestionRepository.java
@@ -1,8 +1,0 @@
-package com.bee.beehomepagebackend.generation.question.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuestionRepository extends JpaRepository<Question, Long> {
-}
