+++ src/main/java/com/bee/beehomepagebackend/recruitment/admin/answer/AnswerAdminController.java
@@ -1,0 +1,4 @@
+package com.bee.beehomepagebackend.recruit.admin.answer;
+
+public class AnswerAdminController {
+}
