+++ src/main/java/com/bee/beehomepagebackend/recruitment/admin/applicant/ApplicantAdminController.java
@@ -1,0 +1,4 @@
+package com.bee.beehomepagebackend.recruit.admin.applicant;
+
+public class ApplicantAdminController {
+}
