+++ src/main/java/com/bee/beehomepagebackend/recruitment/answer/domain/Answer.java
@@ -1,0 +1,40 @@
+package com.bee.beehomepagebackend.recruit.answer.domain;
+
+import com.bee.beehomepagebackend.recruit.Recruit;
+import com.bee.beehomepagebackend.recruit.applicant.domain.Applicant;
+import com.bee.beehomepagebackend.recruit.question.domain.Question;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+
+import java.time.LocalDateTime;
+
+@Getter
+@NoArgsConstructor
+@Entity
+@Table(name = "answer")
+public class Answer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "applicant_id")
+    private Applicant applicant;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "question_id")
+    private Question question;
+
+    private String content;
+
+    @CreatedDate
+    private LocalDateTime createdDateTime;
+
+    @LastModifiedDate
+    private LocalDateTime modifiedDateTime;
+
+}
