+++ src/main/java/com/bee/beehomepagebackend/recruitment/answer/domain/AnswerRepository.java
@@ -1,0 +1,9 @@
+package com.bee.beehomepagebackend.recruit.answer.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AnswerRepository extends JpaRepository<Answer, Long> {
+
+}
