+++ src/main/java/com/bee/beehomepagebackend/recruitment/applicant/domain/ApplicantRepository.java
@@ -1,0 +1,8 @@
+package com.bee.beehomepagebackend.recruit.applicant.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ApplicantRepository extends JpaRepository<Applicant, Long> {
+}
