+++ src/main/java/com/bee/beehomepagebackend/recruitment/applicant/service/ApplicantService.java
@@ -1,0 +1,18 @@
+package com.bee.beehomepagebackend.recruit.applicant.service;
+
+import com.bee.beehomepagebackend.recruit.applicant.domain.ApplicantRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class ApplicantService {
+
+    private final ApplicantRepository applicantRepository;
+
+    public void createApplicant() {
+
+    }
+}
