+++ src/main/java/com/bee/beehomepagebackend/recruitment/applicant/service/request/CreateApplicantServiceRequest.java
@@ -1,0 +1,9 @@
+package com.bee.beehomepagebackend.recruit.applicant.service.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class CreateApplicantServiceRequest {
+}
