+++ src/main/java/com/bee/beehomepagebackend/recruitment/enums/Generation.java
@@ -1,0 +1,8 @@
+package com.bee.beehomepagebackend.recruit.enums;
+
+public enum Generation {
+
+    FIRST,
+    SECOND,
+    THIRD
+}
