+++ src/main/java/com/bee/beehomepagebackend/recruitment/enums/Position.java
@@ -1,0 +1,9 @@
+package com.bee.beehomepagebackend.recruit.enums;
+
+public enum Position {
+
+    BACKEND_DEVELOPER,
+    FRONTEND_DEVELOPER,
+    DESIGNER;
+
+}
