+++ src/main/java/com/bee/beehomepagebackend/recruitment/question/controller/response/GetAllQuestionsResponse.java
@@ -1,0 +1,20 @@
+package com.bee.beehomepagebackend.recruit.question.controller.response;
+
+import com.bee.beehomepagebackend.recruit.question.domain.Question;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@NoArgsConstructor
+public class GetAllQuestionsResponse {
+
+    private List<Question> questions;
+
+    @Builder
+    public GetAllQuestionsResponse(List<Question> questions) {
+        this.questions = questions;
+    }
+}
