+++ src/main/java/com/bee/beehomepagebackend/recruitment/question/domain/QuestionRepository.java
@@ -1,0 +1,8 @@
+package com.bee.beehomepagebackend.recruit.question.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface QuestionRepository extends JpaRepository<Question, Long> {
+}
