+++ src/main/java/com/bee/beehomepagebackend/recruitment/recruit/RecruitController.java
@@ -1,0 +1,18 @@
+package com.bee.beehomepagebackend.recruitment;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/recruits")
+public class RecruitController {
+
+    @PostMapping("/apply")
+    public ResponseEntity<ApplyRecruitRequest> applyRecruit() {
+        return ResponseEntity.ok(null);
+    }
+
+
+}
