+++ src/main/java/com/bee/beehomepagebackend/recruitment/recruit/RecruitRepository.java
@@ -1,0 +1,8 @@
+package com.bee.beehomepagebackend.recruitment;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RecruitRepository extends JpaRepository<Recruit, Integer> {
+}
