+++ src/main/java/com/bee/beehomepagebackend/recruitment/recruit/RecruitService.java
@@ -1,0 +1,7 @@
+package com.bee.beehomepagebackend.recruitment;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class RecruitService {
+}
